requirements.txt
# python 3.6.6
certifi==2018.8.13
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.0.1
django-reversion==3.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.5.5
Pillow==5.2.0
pycryptodome==3.6.6
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
qrcode==6.0
requests==2.19.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
